Group the task to process cogs in seperated nodes
